Extract networking constants
