+++ tests/Jellyfin.Networking.Tests/NetworkParseTests.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using Jellyfin.Networking.Configuration;
+using Jellyfin.Networking.Extensions;
 using Jellyfin.Networking.Manager;
 using MediaBrowser.Common.Configuration;
-using MediaBrowser.Common.Net;
 using MediaBrowser.Model.Net;
 using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
